requirements.txt
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
pandas
strands-agents
strands-agents-tools
mcp
awslabs-aws-documentation-mcp-server>=0.1.4